chore(test): move non-Client specific test helpers to katapult_test.go
