+++ pkg/katapult/client_test.go
@@ -1,24 +1,18 @@
 package katapult
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"os"
-	"reflect"
 	"testing"
 	"time"
 
-	"github.com/augurysys/timestamp"
 	"github.com/krystal/go-katapult/internal/codec"
-	"github.com/krystal/go-katapult/internal/golden"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -84,66 +78,9 @@ var (
 	}
 )
 
-func timestampPtr(unixtime int64) *timestamp.Timestamp {
-	ts := timestamp.Timestamp(time.Unix(unixtime, 0).UTC())
-
-	return &ts
-}
-
-func strictUmarshal(r io.Reader, v interface{}) error {
-	d := json.NewDecoder(r)
-	d.DisallowUnknownFields()
-
-	return d.Decode(v)
-}
-
-func fixture(name string) []byte {
-	file := fmt.Sprintf("fixtures/%s.json", name)
-	c, err := ioutil.ReadFile(file)
-	if err != nil {
-		panic(err)
-	}
-
-	return c
-}
-
-func testJSONMarshaling(t *testing.T, input interface{}) {
-	c := &codec.JSON{}
-
-	buf := &bytes.Buffer{}
-	err := c.Encode(input, buf)
-	require.NoError(t, err, "encoding failed")
-
-	if golden.Update() {
-		golden.Set(t, buf.Bytes())
-	}
-
-	g := golden.Get(t)
-	assert.Equal(t, string(g), buf.String(), "encoding does not match golden")
-
-	got := reflect.New(reflect.TypeOf(input).Elem()).Interface()
-	err = c.Decode(bytes.NewBuffer(g), got)
-	require.NoError(t, err, "decoding golden failed")
-	assert.Equal(t, input, got,
-		"decoding from golden does not match expected object",
-	)
-}
-
-func testQueryableEncoding(t *testing.T, obj queryable) {
-	qs := obj.queryValues()
-	queryStr := qs.Encode()
-
-	if golden.Update() {
-		golden.Set(t, []byte(queryStr))
-	}
-
-	g := string(golden.Get(t))
-	assert.Equal(t, queryStr, g, "query string does not match golden")
-
-	parsedQuery, err := url.ParseQuery(g)
-	require.NoError(t, err, "parsing golden query string failed")
-	assert.Equal(t, qs, &parsedQuery, "parsed golden values do not match")
-}
+//
+// Helpers
+//
 
 func assertFieldSpec(t *testing.T, r *http.Request, spec string) {
 	assert.Equal(t, spec, r.Header.Get("X-Field-Spec"))
@@ -166,6 +103,10 @@ func assertAuthorization(t *testing.T, r *http.Request) {
 var (
 	testAPIKey    = "9d7831d8-03f1-4b4c-a1c3-97272ddefe6a"
 	testUserAgent = "go-katapult/test"
+
+	testDefaultBaseURL   = &url.URL{Scheme: "https", Host: "api.katapult.io"}
+	testDefaultUserAgent = "go-katapult"
+	testDefaultTimeout   = time.Second * 60
 )
 
 // prepareTestClient creates a test HTTP server for mock API responses, and
@@ -206,11 +147,9 @@ func prepareTestClient() (
 	return client, mux, url.String(), server.Close
 }
 
-var (
-	testDefaultBaseURL   = &url.URL{Scheme: "https", Host: "api.katapult.io"}
-	testDefaultUserAgent = "go-katapult"
-	testDefaultTimeout   = time.Second * 60
-)
+//
+// Tests
+//
 
 func TestNewClient(t *testing.T) {
 	type args struct {
